Moved IndexedDB utils into their own folder
