+++ app/assets/js/src/persist/idb.ts
@@ -1,5 +1,9 @@
 import type { PersistApi, PersistOptions } from 'persist/PersistApi';
-import type { DefaultsFor } from 'utils';
+import {
+	ObjectStoreName,
+	doDatabaseTransaction,
+	type DefaultsFor,
+} from 'utils';
 import { bake } from './bake';
 
 const defaultOptions = {
@@ -17,60 +21,6 @@ function getStorageKey(key: string, options: Required<PersistOptions>): string {
 	}
 }
 
-const dbName = 'orange-twist';
-const objectStoreName = 'orange-twist';
-
-let db: IDBDatabase | null = null;
-/**
- * Get a handle to the database, opening it if it wasn't
- * already open.
- */
-function getDatabase(): Promise<IDBDatabase> {
-	return new Promise((resolve, reject) => {
-		if (db) {
-			resolve(db);
-			return;
-		}
-
-		const request = indexedDB.open(dbName);
-
-		// Create object store when database is first created
-		request.addEventListener('upgradeneeded', () => {
-			request.result.createObjectStore(objectStoreName);
-		});
-
-		// Handle success
-		request.addEventListener('success', () => {
-			db = request.result;
-			resolve(request.result);
-		});
-
-		// Handle errors
-		request.addEventListener('error', reject);
-		request.addEventListener('blocked', reject);
-	});
-}
-
-/**
- * Handle the repetitive paperwork like starting a transaction
- * and wrapping things in a Promise.
- */
-function doTransaction(
-	mode: IDBTransactionMode,
-	callback: (objectStore: IDBObjectStore) => IDBRequest
-): Promise<unknown> {
-	return new Promise((resolve, reject) => {
-		getDatabase().then((db) => {
-			const transaction = db.transaction(objectStoreName, mode);
-			const objectStore = transaction.objectStore(objectStoreName);
-			const request = callback(objectStore);
-
-			request.addEventListener('success', () => resolve(request.result));
-			request.addEventListener('error', () => reject(request.error));
-		}).catch(reject);
-	});
-}
-
 /**
  * A {@linkcode PersistApi} interface for working with the
  * IndexedDB API.
@@ -83,8 +33,9 @@ export const idb: PersistApi = {
 		};
 		const storageKey = getStorageKey(key, fullOptions);
 
-		await doTransaction(
+		await doDatabaseTransaction(
 			'readwrite',
+			ObjectStoreName.DATA,
 			(objectStore) => objectStore.put(data, storageKey)
 		);
 	},
@@ -96,8 +47,9 @@ export const idb: PersistApi = {
 		};
 		const storageKey = getStorageKey(key, fullOptions);
 
-		return doTransaction(
+		return doDatabaseTransaction(
 			'readonly',
+			ObjectStoreName.DATA,
 			(objectStore) => objectStore.get(storageKey)
 		);
 	},
@@ -109,8 +61,9 @@ export const idb: PersistApi = {
 		};
 		const storageKey = getStorageKey(key, fullOptions);
 
-		await doTransaction(
+		await doDatabaseTransaction(
 			'readwrite',
+			ObjectStoreName.DATA,
 			(objectStore) => objectStore.delete(storageKey)
 		);
 	},
